Refs LOD-23 and LOD-33.

+++ tests/com/iinteractive/packager/processes/FolderCreationProcessTests.groovy
@@ -14,24 +14,28 @@ public class FolderCreationProcessTests {
 
 	@Before
 	public void cleanUp() {
-		File baseDir = new File("C:\\JUnitTestFolders");
-		File[] files = baseDir.listFiles();
-		for(int i = 0; i < files.length; i++) {
-			DeleteDirectory(files[i]);
+		File baseDir = new File("JUnitTestFolders");
+		if(baseDir.isDirectory()) {
+			File[] files = baseDir.listFiles();
+			for(int i = 0; i < files.length; i++) {
+				DeleteDirectory(files[i]);
+			}
 		}
 		
-		new File("C:\\JUnitTestFolders\\testTwoD1").mkdir();
-		new File("C:\\JUnitTestFolders\\testThreeD1").mkdir();
-		new File("C:\\JUnitTestFolders\\testThreeD1\\testThreeD2").mkdir();
-		new File("C:\\JUnitTestFolders\\testFiveD1").mkdir();
+		assert new File("JUnitTestFolders\\testTwoD1").mkdirs();
+		assert new File("JUnitTestFolders\\testThreeD1").mkdirs();
+		assert new File("JUnitTestFolders\\testThreeD1\\testThreeD2").mkdirs();
+		assert new File("JUnitTestFolders\\testFiveD1").mkdirs();
 	}
 	
 	@After
 	public void cleanUp2() {
-		File baseDir = new File("C:\\JUnitTestFolders");
-		File[] files = baseDir.listFiles();
-		for(int i = 0; i < files.length; i++) {
-			DeleteDirectory(files[i]);
+		File baseDir = new File("JUnitTestFolders");
+		if(baseDir.isDirectory()) {
+			File[] files = baseDir.listFiles();
+			for(int i = 0; i < files.length; i++) {
+				DeleteDirectory(files[i]);
+			}
 		}
 	}
 	
@@ -51,7 +55,7 @@ public class FolderCreationProcessTests {
 	 */
 	@Test
 	public void testOneCreateFolder() {
-		File file = new File("C:\\JUnitTestFolders\\testOneD1\\testOneD2\\file.txt");
+		File file = new File("JUnitTestFolders\\testOneD1\\testOneD2\\file.txt");
 		boolean pass = true;
 		
 		pass = pass && !file.exists();
@@ -74,7 +78,7 @@ public class FolderCreationProcessTests {
 	 */
 	@Test
 	public void testTwoCreateFolder() {
-		File file = new File("C:\\JUnitTestFolders\\testTwoD1\\testTwoD2\\file.txt");
+		File file = new File("JUnitTestFolders\\testTwoD1\\testTwoD2\\file.txt");
 		boolean pass = true;
 		
 		pass = pass && !file.exists();
@@ -97,7 +101,7 @@ public class FolderCreationProcessTests {
 	 */
 	@Test
 	public void testThreeCreateFolder() {
-		File file = new File("C:\\JUnitTestFolders\\testThreeD1\\testThreeD2\\file.txt");
+		File file = new File("JUnitTestFolders\\testThreeD1\\testThreeD2\\file.txt");
 		boolean pass = true;
 		
 		pass = pass && !file.exists();
@@ -120,7 +124,7 @@ public class FolderCreationProcessTests {
 	 */
 	@Test
 	public void testFourCreateFolder() {
-		File file = new File("C:\\JUnitTestFolders\\testFourD1\\file.txt");
+		File file = new File("JUnitTestFolders\\testFourD1\\file.txt");
 		boolean pass = true;
 		
 		pass = pass && !file.exists();
@@ -140,7 +144,7 @@ public class FolderCreationProcessTests {
 	 */
 	@Test
 	public void testFiveCreateFolder() {
-		File file = new File("C:\\JUnitTestFolders\\testFiveD1\\file.txt");
+		File file = new File("JUnitTestFolders\\testFiveD1\\file.txt");
 		boolean pass = true;
 		
 		pass = pass && !file.exists();
